+++ previous-versions/accounting/src/sdk/models/operations/createbanktransactions.ts
@@ -11,25 +11,40 @@ export class CreateBankTransactionsRequest extends SpeakeasyBase {
     createBankTransactions?: shared.CreateBankTransactions;
 
     /**
-     * Unique identifier for an account
+     * Unique identifier for an account.
      */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=accountId" })
     accountId: string;
 
+    /**
+     * Allow a sync upon push completion.
+     */
     @SpeakeasyMetadata({ data: "queryParam, style=form;explode=true;name=allowSyncOnPushComplete" })
     allowSyncOnPushComplete?: boolean;
 
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
+    /**
+     * Unique identifier for a connection.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=connectionId" })
     connectionId: string;
 
+    /**
+     * Time limit for the push operation to complete before it is timed out.
+     */
     @SpeakeasyMetadata({ data: "queryParam, style=form;explode=true;name=timeoutInMinutes" })
     timeoutInMinutes?: number;
 }
 
 export class CreateBankTransactionsResponse extends SpeakeasyBase {
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -45,9 +60,15 @@ export class CreateBankTransactionsResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
 }