+++ previous-versions/accounting/src/sdk/models/operations/getbankaccount.ts
@@ -8,14 +8,20 @@ import { AxiosResponse } from "axios";
 
 export class GetBankAccountRequest extends SpeakeasyBase {
     /**
-     * Unique identifier for an account
+     * Unique identifier for an account.
      */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=accountId" })
     accountId: string;
 
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
+    /**
+     * Unique identifier for a connection.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=connectionId" })
     connectionId: string;
 }
@@ -27,6 +33,9 @@ export class GetBankAccountResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     bankAccount?: shared.BankAccount;
 
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -36,9 +45,15 @@ export class GetBankAccountResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
 }