+++ previous-versions/accounting/src/sdk/models/operations/getbillpayments.ts
@@ -7,9 +7,15 @@ import * as shared from "../shared";
 import { AxiosResponse } from "axios";
 
 export class GetBillPaymentsRequest extends SpeakeasyBase {
+    /**
+     * Unique identifier for a bill payment.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=billPaymentId" })
     billPaymentId: string;
 
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 }
@@ -21,6 +27,9 @@ export class GetBillPaymentsResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     billPayment?: shared.BillPayment;
 
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -30,9 +39,15 @@ export class GetBillPaymentsResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
 }