+++ previous-versions/accounting/src/sdk/models/shared/banktransactiontype.ts
@@ -3,7 +3,7 @@
  */
 
 /**
- * Type of transaction for the bank statement line
+ * Type of transaction for the bank statement line.
  */
 export enum BankTransactionType {
     Unknown = "Unknown",