+++ previous-versions/accounting/src/sdk/sdk.ts
@@ -77,8 +77,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.37.0";
-    genVersion = "2.109.3";
+    sdkVersion = "0.38.0";
+    genVersion = "2.144.7";
+    userAgent = "speakeasy-sdk/typescript 0.38.0 2.144.7 3.0.0 @codat/accounting";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);